Refactored some methods from Series Controller and Library Controller to a new Search Controller to keep things organized
