+++ UI/Web/src/app/_services/library.service.ts
@@ -113,12 +113,4 @@ export class LibraryService {
       return this.libraryTypes[libraryId];
     }));
   }
-
-  search(term: string) {
-    if (term === '') {
-      return of(new SearchResultGroup());
-    }
-    return this.httpClient.get<SearchResultGroup>(this.baseUrl + 'library/search?queryString=' + encodeURIComponent(term));
-  }
-
 }
