+++ UI/Web/src/app/_services/series.service.ts
@@ -78,14 +78,6 @@ export class SeriesService {
     return this.httpClient.get<ChapterMetadata>(this.baseUrl + 'series/chapter-metadata?chapterId=' + chapterId);
   }
 
-  getSeriesForMangaFile(mangaFileId: number) {
-    return this.httpClient.get<Series | null>(this.baseUrl + 'series/series-for-mangafile?mangaFileId=' + mangaFileId);
-  }
-
-  getSeriesForChapter(chapterId: number) {
-    return this.httpClient.get<Series | null>(this.baseUrl + 'series/series-for-chapter?chapterId=' + chapterId);
-  }
-
   delete(seriesId: number) {
     return this.httpClient.delete<boolean>(this.baseUrl + 'series/' + seriesId);
   }
