+++ UI/Web/src/app/nav/nav-header/nav-header.component.ts
@@ -6,7 +6,7 @@ import { debounceTime, distinctUntilChanged, filter, takeUntil, tap } from 'rxjs
 import { Chapter } from 'src/app/_models/chapter';
 import { MangaFile } from 'src/app/_models/manga-file';
 import { ScrollService } from 'src/app/_services/scroll.service';
-import { SeriesService } from 'src/app/_services/series.service';
+import { SearchService } from 'src/app/_services/search.service';
 import { FilterQueryParam } from '../../shared/_services/filter-utilities.service';
 import { CollectionTag } from '../../_models/collection-tag';
 import { Library } from '../../_models/library';
@@ -52,8 +52,8 @@ export class NavHeaderComponent implements OnInit, OnDestroy {
   private readonly onDestroy = new Subject<void>();
 
   constructor(public accountService: AccountService, private router: Router, public navService: NavService,
-    private libraryService: LibraryService, public imageService: ImageService, @Inject(DOCUMENT) private document: Document,
-    private scrollService: ScrollService, private seriesService: SeriesService, private readonly cdRef: ChangeDetectorRef) {
+    public imageService: ImageService, @Inject(DOCUMENT) private document: Document,
+    private scrollService: ScrollService, private searchService: SearchService, private readonly cdRef: ChangeDetectorRef) {
       this.scrollElem = this.document.body;
     }
 
@@ -110,7 +110,7 @@ export class NavHeaderComponent implements OnInit, OnDestroy {
       this.searchTerm = val.trim();
       this.cdRef.markForCheck();
 
-      this.libraryService.search(val.trim()).pipe(takeUntil(this.onDestroy)).subscribe(results => {
+      this.searchService.search(val.trim()).pipe(takeUntil(this.onDestroy)).subscribe(results => {
         this.searchResults = results;
         this.isLoading = false;
         this.cdRef.markForCheck();
@@ -185,7 +185,7 @@ export class NavHeaderComponent implements OnInit, OnDestroy {
 
   clickFileSearchResult(item: MangaFile) {
     this.clearSearch();
-    this.seriesService.getSeriesForMangaFile(item.id).subscribe(series => {
+    this.searchService.getSeriesForMangaFile(item.id).subscribe(series => {
       if (series !== undefined && series !== null) {
         this.router.navigate(['library', series.libraryId, 'series', series.id]);
       }
@@ -194,7 +194,7 @@ export class NavHeaderComponent implements OnInit, OnDestroy {
 
   clickChapterSearchResult(item: Chapter) {
     this.clearSearch();
-    this.seriesService.getSeriesForChapter(item.id).subscribe(series => {
+    this.searchService.getSeriesForChapter(item.id).subscribe(series => {
       if (series !== undefined && series !== null) {
         this.router.navigate(['library', series.libraryId, 'series', series.id]);
       }
